+++ src/main/java/com/chs/cafeapp/exception/type/ErrorCode.java
@@ -13,15 +13,18 @@ public enum ErrorCode {
   ALREADY_EMAIL_AUTH_MEMBER(400, "이미 이메일 인증을 완료한 사용자입니다. 로그인해서 서비스를 확인하세요."),
   EXPIRED_DATE_TIME_FOR_EMAIL_AUTH(400, "이메일 인증은 메일 전송 24시간 이내에 해주셔야 합니다. 이메일을 다시 보냈습니다. 인증을 24시간 이내에 해주십시오."),
   NOT_EMAIL_AUTH(400, "이메일 인증 후 로그인이 가능합니다."),
-  NOT_EXISTS_MEMBER_LOGIN_ID(400, "로그인 아이디가 존재하지 않습니다. 다시 확인해 주십시오."),
-  NOT_MATCH_MEMBER_PASSWORD(400, "로그인 비밀번호가 틀렸습니다."),
+  NOT_EXISTS_LOGIN_ID(400, "로그인 아이디가 존재하지 않습니다. 다시 확인해 주십시오."),
+  NOT_MATCH_PASSWORD(400, "로그인 비밀번호가 틀렸습니다."),
   NOT_MATCH_ORIGIN_PASSWORD(400, "기존 비밀번호가 맞지 않습니다. 다시 확인해 주십시오."),
+  NOT_EQUALS_PASSWORD_REPASSWORD(400, "비밀번호와 비밀번호확인이 같지 않습니다. 다시 확인해 주십시오."),
   NOT_VALID_REFRESH_TOKEN(400, "Refresh Token이 유효하지 않습니다."),
   NOTING_ACCESS_TOKEN(400, "Access Token이 없습니다."),
   NOT_MATCH_REFRESH_TOKEN_MEMBER(400, "토큰의 유저 정보가 일치하지 않습니다."),
   UN_AUTHORIZATION(401, "인증이 되지 않았습니다."),
   NOT_MATCH_AUTHORIZATION(403, "접근 권한이 올바르지 않습니다."),
   LOGOUT_MEMBER(400, "로그아웃한 사용자입니다."),
+  NOT_MATCH_MEMBER_ROLE(400, "카페 사용자의 권한이 아닙니다."),
+  NOT_MATCH_ADMIN_ROLE(400, "카페 관리자의 권한이 아닙니다."),
   NOT_MATCH_ROLE(400, "권한(Role)이 동일하지 않습니다."),
   TOO_MANY_ROLE(400, "권한(Role)은 각 1개씩만 존재해야합니다."),
   INVALID_ACCESS_TOKEN(400, "유효하지 않은 Access Token 값입니다."),
